feat: Separate pages
I'm adding links to Comparison & Browse in the Home
